fix: undo identifier struct type changes (#328)

+++ packages/contracts-bedrock/interfaces/L2/IL2ToL2CrossDomainMessenger.sol
@@ -3,9 +3,9 @@ pragma solidity ^0.8.0;
 
 struct Identifier {
     address origin;
-    uint64 blockNumber;
-    uint32 logIndex;
-    uint64 timestamp;
+    uint256 blockNumber;
+    uint256 logIndex;
+    uint256 timestamp;
     uint256 chainId;
 }
